Revert e5ba8616
