+++ .env.example
@@ -77,7 +77,7 @@ NEXT_PUBLIC_HELPSCOUT_KEY=
 SEND_FEEDBACK_EMAIL=
 
 # Sengrid
-# Used for email reminders in workflows and internal sync services
+# Used for email reminders in workflows
 SENDGRID_API_KEY=
 SENDGRID_EMAIL=
 
@@ -133,9 +133,6 @@ NEXT_PUBLIC_TEAM_IMPERSONATION=false
 
 # Close.com internal CRM
 CLOSECOM_API_KEY=
-
-# Sendgrid internal sync service
-SENDGRID_SYNC_API_KEY=
 
 # Sendgrid internal email sender
 SENDGRID_API_KEY=
