+++ packages/app-store/_pages/setup/index.tsx
@@ -10,7 +10,6 @@ export const AppSetupMap = {
   "caldav-calendar": dynamic(() => import("../../caldavcalendar/pages/setup")),
   zapier: dynamic(() => import("../../zapier/pages/setup")),
   closecom: dynamic(() => import("../../closecomothercalendar/pages/setup")),
-  sendgrid: dynamic(() => import("../../sendgridothercalendar/pages/setup")),
 };
 
 export const AppSetupPage = (props: { slug: string }) => {
