+++ packages/app-store/apps.metadata.generated.ts
@@ -30,7 +30,6 @@ import { metadata as qr_code_meta } from "./qr_code/_metadata";
 import { metadata as rainbow_meta } from "./rainbow/_metadata";
 import { metadata as raycast_meta } from "./raycast/_metadata";
 import { metadata as riverside_meta } from "./riverside/_metadata";
-import { metadata as sendgridothercalendar_meta } from "./sendgridothercalendar/_metadata";
 import { metadata as sirius_video_meta } from "./sirius_video/_metadata";
 import { metadata as stripepayment_meta } from "./stripepayment/_metadata";
 import { metadata as tandemvideo_meta } from "./tandemvideo/_metadata";
@@ -71,7 +70,6 @@ export const appStoreMetadata = {
   rainbow: rainbow_meta,
   raycast: raycast_meta,
   riverside: riverside_meta,
-  sendgridothercalendar: sendgridothercalendar_meta,
   sirius_video: sirius_video_meta,
   stripepayment: stripepayment_meta,
   tandemvideo: tandemvideo_meta,
