+++ packages/app-store/apps.server.generated.ts
@@ -29,7 +29,6 @@ export const apiHandlers = {
   rainbow: import("./rainbow/api"),
   raycast: import("./raycast/api"),
   riverside: import("./riverside/api"),
-  sendgridothercalendar: import("./sendgridothercalendar/api"),
   sirius_video: import("./sirius_video/api"),
   stripepayment: import("./stripepayment/api"),
   tandemvideo: import("./tandemvideo/api"),
