+++ packages/app-store/closecomothercalendar/test/globals.ts
@@ -1,11 +1,8 @@
 jest.mock("@calcom/lib/logger", () => ({
-  default: {
-    getChildLogger: () => ({
-      debug: jest.fn(),
-      error: jest.fn(),
-      log: jest.fn(),
-    }),
-  },
+  debug: jest.fn(),
+  error: jest.fn(),
+  log: jest.fn(),
+  getChildLogger: jest.fn(),
 }));
 
 jest.mock("@calcom/lib/crypto", () => ({
