+++ packages/app-store/closecomothercalendar/test/lib/CalendarService.test.ts
@@ -1,5 +1,19 @@
 import CloseCom from "@calcom/lib/CloseCom";
+import {
+  getCloseComContactIds,
+  getCustomActivityTypeInstanceData,
+  getCloseComCustomActivityTypeFieldsIds,
+  getCloseComLeadId,
+} from "@calcom/lib/CloseComeUtils";
 import type { CalendarEvent } from "@calcom/types/Calendar";
+
+jest.mock("@calcom/lib/CloseCom", () => ({
+  default: class {
+    constructor() {
+      /* Mock */
+    }
+  },
+}));
 
 afterEach(() => {
   jest.resetAllMocks();
@@ -12,8 +26,9 @@ test("check generic lead generator: already exists", async () => {
       data: [{ name: "From Cal.com", id: "abc" }],
     }),
   } as any;
-  debugger;
-  const id = await CloseCom.prototype.getCloseComLeadId();
+
+  const closeCom = new CloseCom("someKey");
+  const id = await getCloseComLeadId(closeCom);
   expect(id).toEqual("abc");
 });
 
@@ -26,7 +41,8 @@ test("check generic lead generator: doesn't exist", async () => {
     create: () => ({ id: "def" }),
   } as any;
 
-  const id = await CloseCom.prototype.getCloseComLeadId();
+  const closeCom = new CloseCom("someKey");
+  const id = await getCloseComLeadId(closeCom);
   expect(id).toEqual("def");
 });
 
@@ -45,7 +61,8 @@ test("retrieve contact IDs: all exist", async () => {
     search: () => ({ data: attendees }),
   } as any;
 
-  const contactIds = await CloseCom.prototype.getCloseComContactIds(event.attendees, "leadId");
+  const closeCom = new CloseCom("someKey");
+  const contactIds = await getCloseComContactIds(event.attendees, closeCom, "leadId");
   expect(contactIds).toEqual(["test1", "test2"]);
 });
 
@@ -62,7 +79,8 @@ test("retrieve contact IDs: some don't exist", async () => {
     create: () => ({ id: "test3" }),
   } as any;
 
-  const contactIds = await CloseCom.prototype.getCloseComContactIds(event.attendees, "leadId");
+  const closeCom = new CloseCom("someKey");
+  const contactIds = await getCloseComContactIds(event.attendees, closeCom, "leadId");
   expect(contactIds).toEqual(["test1", "test3"]);
 });
 
@@ -87,11 +105,15 @@ test("retrieve custom fields for custom activity type: type doesn't exist, no fi
     },
   } as any;
 
-  const contactIds = await CloseCom.prototype.getCloseComCustomActivityTypeFieldsIds([
-    ["Attendees", "", true, true],
-    ["Date & Time", "", true, true],
-    ["Time Zone", "", true, true],
-  ]);
+  const closeCom = new CloseCom("someKey");
+  const contactIds = await getCloseComCustomActivityTypeFieldsIds(
+    [
+      ["Attendees", "", true, true],
+      ["Date & Time", "", true, true],
+      ["Time Zone", "", true, true],
+    ],
+    closeCom
+  );
   expect(contactIds).toEqual({
     activityType: "type1",
     fields: ["field9A", "field11D", "field9T"],
@@ -119,11 +141,15 @@ test("retrieve custom fields for custom activity type: type exists, no field cre
     },
   } as any;
 
-  const contactIds = await CloseCom.prototype.getCloseComCustomActivityTypeFieldsIds([
-    ["Attendees", "", true, true],
-    ["Date & Time", "", true, true],
-    ["Time Zone", "", true, true],
-  ]);
+  const closeCom = new CloseCom("someKey");
+  const contactIds = await getCloseComCustomActivityTypeFieldsIds(
+    [
+      ["Attendees", "", true, true],
+      ["Date & Time", "", true, true],
+      ["Time Zone", "", true, true],
+    ],
+    closeCom
+  );
   expect(contactIds).toEqual({
     activityType: "typeX",
     fields: ["fieldY", "field11D", "field9T"],
@@ -142,7 +168,7 @@ test("prepare data to create custom activity type instance: two attendees, no ad
   const event = {
     attendees,
     startTime: now.toISOString(),
-  } as unknown as CalendarEvent;
+  } as CalendarEvent;
 
   CloseCom.prototype.activity = {
     type: {
@@ -170,11 +196,16 @@ test("prepare data to create custom activity type instance: two attendees, no ad
     create: () => ({ id: "def" }),
   } as any;
 
-  const data = await CloseCom.prototype.getCustomActivityTypeInstanceData(event, [
-    ["Attendees", "", true, true],
-    ["Date & Time", "", true, true],
-    ["Time Zone", "", true, true],
-  ]);
+  const closeCom = new CloseCom("someKey");
+  const data = await getCustomActivityTypeInstanceData(
+    event,
+    [
+      ["Attendees", "", true, true],
+      ["Date & Time", "", true, true],
+      ["Time Zone", "", true, true],
+    ],
+    closeCom
+  );
   expect(data).toEqual({
     custom_activity_type_id: "type1",
     lead_id: "def",
@@ -221,11 +252,16 @@ test("prepare data to create custom activity type instance: one attendees, with 
     }),
   } as any;
 
-  const data = await CloseCom.prototype.getCustomActivityTypeInstanceData(event, [
-    ["Attendees", "", true, true],
-    ["Date & Time", "", true, true],
-    ["Time Zone", "", true, true],
-  ]);
+  const closeCom = new CloseCom("someKey");
+  const data = await getCustomActivityTypeInstanceData(
+    event,
+    [
+      ["Attendees", "", true, true],
+      ["Date & Time", "", true, true],
+      ["Time Zone", "", true, true],
+    ],
+    closeCom
+  );
   expect(data).toEqual({
     custom_activity_type_id: "type1",
     lead_id: "abc",
