--- packages/app-store/sendgridothercalendar/README.mdx
@@ -1,10 +1,0 @@
----
-description: SendGrid delivers your transactional and marketing emails through the world's largest cloud-based email delivery platform.
-items:
-  - /api/app-store/sendgridothercalendar/1.png
----
-
-{description}
-
-Features:
- - Creates event attendees as contacts in Sendgrid
