--- packages/app-store/sendgridothercalendar/_metadata.ts
@@ -1,10 +1,0 @@
-import type { AppMeta } from "@calcom/types/App";
-
-import config from "./config.json";
-
-export const metadata = {
-  category: "other",
-  ...config,
-} as AppMeta;
-
-export default metadata;
