--- packages/app-store/sendgridothercalendar/api/add.ts
@@ -1,6 +1,0 @@
-import { defaultHandler } from "@calcom/lib/server";
-
-export default defaultHandler({
-  GET: import("./_getAdd"),
-  POST: import("./_postAdd"),
-});
