--- packages/app-store/sendgridothercalendar/api/check.ts
@@ -1,5 +1,0 @@
-import { defaultHandler } from "@calcom/lib/server";
-
-export default defaultHandler({
-  POST: import("./_postCheck"),
-});
