--- packages/app-store/sendgridothercalendar/config.json
@@ -1,16 +1,0 @@
-{
-  "/*": "Don't modify slug - If required, do it using cli edit command",
-  "name": "Sendgrid",
-  "slug": "sendgrid",
-  "type": "sendgrid_other_calendar",
-  "imageSrc": "/api/app-store/sendgridothercalendar/icon.png",
-  "logo": "/api/app-store/sendgridothercalendar/icon.png",
-  "url": "https://cal.com/apps/sendgrid",
-  "variant": "other_calendar",
-  "categories": ["other"],
-  "publisher": "Cal.com",
-  "email": "help@cal.com",
-  "description": "SendGrid delivers your transactional and marketing emails through the world's largest cloud-based email delivery platform.",
-  "extendsFeature": "User",
-  "__createdUsingCli": true
-}
