--- packages/app-store/sendgridothercalendar/index.ts
@@ -1,3 +1,0 @@
-export * as api from "./api";
-export * as lib from "./lib";
-export { metadata } from "./_metadata";
