--- packages/app-store/sendgridothercalendar/lib/index.ts
@@ -1,1 +1,0 @@
-export { default as CalendarService } from "./CalendarService";
