+++ packages/prisma/seed-app-store.config.json
@@ -101,11 +101,5 @@
     "categories": ["video"],
     "slug": "sirius_video",
     "type": "sirius_video_video"
-  },
-  {
-    "dirName": "sendgrid",
-    "categories": ["other"],
-    "slug": "sendgrid",
-    "type": "sendgrid_other_calendar"
   }
 ]
