+++ turbo.json
@@ -182,7 +182,6 @@
     "$CI",
     "$CLOSECOM_API_KEY",
     "$SENDGRID_API_KEY",
-    "$SENDGRID_SYNC_API_KEY",
     "$SENDGRID_EMAIL",
     "$CRON_API_KEY",
     "$DAILY_API_KEY",
